Fixed payment collection when staking NFTs into session
Payment is only collected conditionally if the payment token is not set
to the 0 address.

Previously, the ERC20 `transferFrom` was not being processed correctly
(by requiring that its return value be `true` - ERC20 transfers can fail
without raising errors according to the standard).

Added 2 tests for payment behavior.

Also added an explicit check that NFTs are being staked by their owners.
Added a test that someone who didn't own an NFT could not stake that NFT
into a Garden of Forking Paths contract.