Updated unstaking behaviour on GOFPFacet
And modified tests so that they test the new behaviour rather than the
old one.

The changes:

1. Players can unstake their NFTs from a session whenever they want to -
even if the session is active.
2. We added a staking guard on each session so that, once an NFT has
been staked into a session, it can never be staked into the same session
again.

This improves the level of decentralization of the Garden of Forking
Paths mechanic - it doesn't hold NFTs hostage to game masters setting
sessions inactive.