Removed slot types
The slot types are implicitly defined by the items that can be equipped
in those slots.

Having an explicit notion of slot type with no conceivable way to enforce
it creates two sources of truth about what a slot does - it's type and
the set of items equippable in that slot. This makes the system more
complex and difficult to work with, with no additional benefit.